requirements.txt
slack-sdk==3.26.1
python-dotenv==1.0.0
schedule==1.2.1
flask==3.0.2 